refactor: move updateTemplate graphql resolver to a separate file
