+++ templates/graphql/template/index.js
@@ -1,0 +1,4 @@
+const { createTemplate } = require("./resolvers/create-template");
+const { updateTemplate } = require("./resolvers/update-template");
+
+module.exports = { createTemplate, updateTemplate };
